feat: add layouts for nested category routing
