+++ apps/web/src/app/layout.tsx
@@ -1,6 +1,7 @@
-import localFont from 'next/font/local';
 import 'ui/styles.css';
 import './globals.css';
+
+import localFont from 'next/font/local';
 
 const pretendard = localFont({
   src: '../../public/font/PretendardVariable.woff2',
@@ -18,7 +19,11 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <body className={pretendard.className}>{children}</body>
+      <body className={pretendard.className}>
+        <main className="flex min-h-screen w-full flex-col items-center ">
+          {children}
+        </main>
+      </body>
     </html>
   );
 }
